Add Environment #define responsible to include the proper instance
