+++ lib/interstate/active_record_class/instance_methods.rb
@@ -1,0 +1,11 @@
+module Interstate
+  module ActiveRecordClass
+    module InstanceMethods
+      def self.included(base)
+        base.after_initialize do |_record|
+          @state_machine = StateMachine.new(self)
+        end
+      end
+    end
+  end
+end
