+++ spec/spec_helper.rb
@@ -2,9 +2,9 @@ require 'simplecov'
 SimpleCov.start
 require 'bundler/setup'
 require 'byebug'
-require 'interstate'
 require 'active_record'
-require_relative '../examples/vehicle/vehicle.rb'
+require 'interstate'
+# require_relative '../examples/vehicle/vehicle.rb'
 
 Dir[File.dirname(__FILE__) + '/models/**/*.rb'].sort.each { |f| require File.expand_path(f) }
